refactor: reorganize project structure by separating pages and components
