+++ src/pages/Project.tsx
@@ -1,12 +1,12 @@
 /* 프로젝트 페이지 */
 
 import React, { useState, useEffect } from 'react';
-import Projects from './Projects.json';
-import NextBtn from '../common/NextBtn';
-import ProjectCard from './ProjectCard';
-import ProjectPreview from './ProjectPreview';
-import { ProjectItem } from '../../types/projects';
-import '../../styles/Project/Project.scss';
+import Projects from '../components/Project/Projects.json';
+import NextBtn from '../components/common/NextBtn';
+import ProjectCard from '../components/Project/ProjectCard';
+import ProjectPreview from '../components/Project/ProjectPreview';
+import { ProjectItem } from '../types/projects';
+import '../styles/Project/Project.scss';
 
 const Project: React.FC = () => {
   // 모달 오픈
