+++ src/pages/Skill.tsx
@@ -1,16 +1,16 @@
 // tech stacks
 
 import React, { useState } from 'react';
-import NextBtn from '../common/NextBtn';
-import Frontend from './Frontend.json';
-import Backend from './Backend.json';
-import Library from './Library.json';
-import Develop from './Develop.json';
-import Others from './Others.json';
-import SkillArea from './SkillArea';
-import { BadgeItem } from '../../types/skills';
+import NextBtn from '../components/common/NextBtn';
+import Frontend from '../components/Skill/Frontend.json';
+import Backend from '../components/Skill/Backend.json';
+import Library from '../components/Skill/Library.json';
+import Develop from '../components/Skill/Develop.json';
+import Others from '../components/Skill/Others.json';
+import SkillArea from '../components/Skill/SkillArea';
+import { BadgeItem } from '../types/skills';
 import FolderIcon from '@mui/icons-material/Folder';
-import '../../styles/Skill/Skill.scss';
+import '../styles/Skill/Skill.scss';
 
 const skillData = [
   { title: 'Frontend', data: Frontend },
